Conflicts:
	application/controllers/Merchant.php
	application/views/all/hotdeal.php

+++ application/views/all/picture.php
@@ -3,18 +3,15 @@
 <script type="text/javascript" src="<?php echo base_url() ?>js/jgrowl/jquery.jgrowl.js"></script>
 <?php echo link_tag('js/jgrowl/jquery.jgrowl.css') ?>
 <script type="text/javascript" src="<?php echo base_url() ?>js/js_custom.js"></script>
-<?php echo link_tag('css/bootstrap.min.css') ?>
 
 <div id='picture-user'>
     <h1><?php echo $page_title; ?></h1>
-    <div id='picture-user-content'>       
-        
+    <div id='picture-user-content'>
         <div id="picture-user-edit-link">
             <?php
             if (check_is_login())
             {
                 $group_id_user = $this->config->item('group_id_user');
-                
                 $login_id = $this->ion_auth->user()->row()->id;
                 $user_allowed_list = $this->m_custom->get_list_of_allow_id('merchant_user_album', 'user_id', $login_id, 'merchant_user_album_id', 'post_type', 'mer');
                 if (check_correct_login_type($group_id_user, $user_allowed_list, $picture_id))
@@ -22,7 +19,6 @@
                     $edit_url = base_url() . "user/edit_merchant_picture/" . $picture_id;
                     echo "<a href='".$edit_url."'>Edit Picture</a>";
                 }
-                
                 $group_id_merchant = $this->config->item('group_id_merchant');
                 $group_id_supervisor = $this->config->item('group_id_supervisor');
                 if (check_correct_login_type($group_id_merchant) || check_correct_login_type($group_id_supervisor))
@@ -32,58 +28,48 @@
                     {                   
                         $merchant_id = $this->ion_auth->user()->row()->su_merchant_id;
                     }
-
                     $merchant_allowed_list = $this->m_custom->get_list_of_allow_id('merchant_user_album', 'merchant_id', $merchant_id, 'merchant_user_album_id', 'post_type', 'mer');
                     $hide_url = base_url() . "merchant/remove_mua_picture/" . $picture_id;
-
                     if (check_allowed_list($merchant_allowed_list, $picture_id))
                     {                       
                         ?>
-            <button type="submit" data-toggle = "modal" data-target = "#myModal_Remove" style="float:right">
-                       Remove Picture
-                    </button>
-                        <div class = "modal fade" id = "myModal_Remove" tabindex = "-1" role = "dialog" aria-labelledby = "myModalLabel" aria-hidden="true">
-   
-   <div class = "modal-dialog" >
-      <div class = "modal-content">
-         
-         <div class = "modal-header">
-            <button type = "button" class = "close" data-dismiss = "modal" aria-hidden = "true">
-                  &times;
-            </button>
-            
-            <h4 class = "modal-title" id = "myModalLabel">
-               Why Need To Remove?
-            </h4>
-         </div>
-         
-         <div class = "modal-body"> 
-             <?php
-                        $action_url = base_url() . "merchant/remove_mua_picture";
-                        $confirm_message = "Confirm that you want to remove this picture that user upload for your company? ";
-                        ?>
-                        <form action="<?php echo $action_url; ?>" onSubmit="return confirm('<?php echo $confirm_message ?>')" method="post" accept-charset="utf-8">
-                            <div id="contact-us-right-form">
-                                        <div style="color:red;font-weight:bold">Please put in the reason to remove this : </div>
-                                        <div id="contact-us-right-form-each">
-                                            <input type="text" placeholder="Remove Reason" id="hide_remark" name="hide_remark"><br/>
+                        <button type="submit" data-toggle = "modal" data-target = "#myModal_Remove" style="float:right">Remove Picture</button>
+                        <div class="modal fade" id="myModal_Remove" tabindex="-1" role="dialog" aria-labelledby="myModalLabel" aria-hidden="true">
+                            <div class="modal-dialog">
+                                <div class="modal-content">
+                                    <div class="modal-header">
+                                        <button type="button" class="bootstrap-close" data-dismiss="modal" aria-hidden="true">&times;</button>
+                                        <h4 class ="modal-title" id="myModalLabel">
+                                            Why Need To Remove?
+                                        </h4>
+                                    </div>
+                                    <div class="modal-body"> 
+                                        <?php
+                                        $action_url = base_url() . "merchant/remove_mua_picture";
+                                        $confirm_message = "Confirm that you want to remove this picture that user upload for your company? ";
+                                        ?>
+                                        <div class="bootstrap-form">
+                                            <form action="<?php echo $action_url; ?>" onSubmit="return confirm('<?php echo $confirm_message ?>')" method="post" accept-charset="utf-8">
+                                                <div class="bootstrap-form-label">
+                                                    <div style="color:red; font-weight:bold">Please put in the reason to remove this :</div>
+                                                </div>
+                                                <div class="bootstrap-form-input">
+                                                    <input type="text" placeholder="Remove Reason" id="hide_remark" name="hide_remark"><br/>
+                                                </div>
+                                                <?php
+                                                echo "<input type='hidden' name='hid_picture_id' value='" . $picture_id . "' />";
+                                                echo "<input type='hidden' name='hid_upload_by_user_id' value='" . $upload_by_user_id . "' />";
+                                                ?>
+                                                <div class="bootstrap-form-button">
+                                                    <button name='button_action' type='submit' value='hide_picture'>Confirm Remove</button>
+                                                </div>
+                                            </form>
                                         </div>
-                                     </div>
+                                    </div>
+                                </div>
+                            </div>
+                        </div>
                         <?php
-                        echo "<input type='hidden' name='hid_picture_id' value='".$picture_id."' />";
-                        echo "<input type='hidden' name='hid_upload_by_user_id' value='".$upload_by_user_id."' />";
-                        echo "<div class = 'modal-footer'>";
-                        echo "<button name='button_action' type='submit' value='hide_picture'>Confirm Remove</button>";
-                        echo "</div>";
-                        echo form_close(); 
-                        ?>
-                          </div>
-         
-      </div><!-- /.modal-content -->
-   </div><!-- /.modal-dialog -->
-  
-</div><!-- /.modal --> 
-    <?php
                     }
                 }
             }
@@ -92,6 +78,17 @@
         <div id="float-fix"></div>
         
         <div id='picture-user-table'>
+            <div id='picture-user-table-row'>
+                <div id='picture-user-table-row-cell' class='picture-user-left-cell'></div>
+                <div id='picture-user-table-row-cell' class='picture-user-center-cell'>
+                    <div id='picture-user-center'>
+                        <div id="picture-user-title">
+                            <a href='<?php echo $merchant_dashboard_url ?>'> <?php echo $merchant_name ?></a>
+                        </div>
+                    </div>
+                </div>
+                <div id='picture-user-table-row-cell' class='picture-user-right-cell'></div>
+            </div>
             <div id='picture-user-table-row'>
                 <div id='picture-user-table-row-cell' class='picture-user-left-cell'>
                     <div id='picture-user-left'>
@@ -109,12 +106,30 @@
                 </div>
                 <div id='picture-user-table-row-cell' class='picture-user-center-cell'>
                     <div id='picture-user-center'>
-                        <div id="picture-user-title">
-                            <a href='<?php echo $merchant_dashboard_url ?>'> <?php echo $merchant_name ?></a>
-                        </div>
                         <div id='picture-user-photo-box'>
                             <img src='<?php echo $image_url ?>'>
                         </div>
+                    </div>
+                </div>
+                <div id='picture-user-table-row-cell' class='picture-user-right-cell'>
+                    <div id='picture-user-right'>
+                        <?php
+                        if (!empty($next_url))
+                        {
+                            ?><a href="<?php echo $next_url ?>"><i class="fa fa-angle-double-right"></i></a><?php
+                        }
+                        else 
+                        {
+                            ?><div id='picture-user-right-gray'><i class="fa fa-angle-double-right"></i></div><?php
+                        }
+                        ?>
+                    </div>
+                </div>
+            </div>
+            <div id='picture-user-table-row'>
+                <div id='picture-user-table-row-cell' class='picture-user-left-cell'></div>
+                <div id='picture-user-table-row-cell' class='picture-user-center-cell'>
+                    <div id='picture-user-center'>
                         <div id="picture-user-sub-title">
                             <?php echo $title ?>
                         </div>
@@ -137,15 +152,20 @@
                                 ?>
                             </div>
                             <div id="picture-user-upload-by">
-                                <?php if(check_is_login()){
+                                <?php 
+                                if(check_is_login())
+                                {
                                     $login_id = $this->ion_auth->user()->row()->id;
-                                    if($picture_user_id != $login_id){
+                                    if($picture_user_id != $login_id)
+                                    {
                                         echo 'Upload by : '.$user_name_url;            
-                                    }                                                     
-                                 }else{
-                                     echo 'Upload by : '.$user_name_url;            
-                                 }
-                                 ?>
+                                    }
+                                }
+                                else
+                                {
+                                    echo 'Upload by : '.$user_name_url;            
+                                }
+                                ?>
                             </div>
                             <div id="float-fix"></div>
                         </div>
@@ -174,22 +194,8 @@
                         </div>
                     </div>
                 </div>
-                <div id='picture-user-table-row-cell' class='picture-user-right-cell'>
-                    <div id='picture-user-right'>
-                        <?php
-                        if (!empty($next_url))
-                        {
-                            ?><a href="<?php echo $next_url ?>"><i class="fa fa-angle-double-right"></i></a><?php
-                        }
-                        else 
-                        {
-                            ?><div id='picture-user-right-gray'><i class="fa fa-angle-double-right"></i></div><?php
-                        }
-                        ?>
-                    </div>
-                </div>
+                <div id='picture-user-table-row-cell' class='picture-user-right-cell'></div>
             </div>
         </div>
-        
     </div>
 </div>
